server/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
python-dotenv==0.19.1